+++ src/infrastructure/commands/generate/index.ts
@@ -1,13 +1,15 @@
 import {loadConfiguration} from '@nestjs/cli/lib/utils/load-configuration';
 import {join, resolve} from 'path';
-import * as lodash from 'lodash';
 import * as fs from 'fs';
 import {CommandUtils} from '@steroidsjs/typeorm/commands/CommandUtils';
-import {Connection} from '@steroidsjs/typeorm';
+import {Connection, DataSource} from '@steroidsjs/typeorm';
 import {format} from '@sqltools/formatter';
+import * as glob from 'glob';
 import {CustomRdbmsSchemaBuilder} from './CustomRdbmsSchemaBuilder';
-import * as glob from "glob";
 import {ModuleHelper} from '../../helpers/ModuleHelper';
+import {getNewPermissions} from '../../utils/getNewPermissions';
+
+const ADD_PERMISSIONS_NAME = 'AddPermissions';
 
 const queryParams = (parameters: any[] | undefined): string => {
     if (!parameters || !parameters.length) {
@@ -15,14 +17,14 @@ const queryParams = (parameters: any[] | undefined): string => {
     }
 
     return `, ${JSON.stringify(parameters)}`;
-}
+};
 
 const prettifyQuery = (query: string) => {
     const formattedQuery = format(query, {indent: '    '});
     query = '\n' + formattedQuery.replace(/^/gm, '            ') + '\n        ';
     query = query.replace(/`/g, '\\`');
     return query;
-}
+};
 
 /**
  * Gets contents of the migration file.
@@ -46,9 +48,62 @@ ${downSqls.join(`
     }
 }
 `;
-}
+};
+
+export const generateMigrationsForPermissions = async (dataSource: DataSource, permissionOptions = {
+    table: 'auth_permission',
+    column: 'name',
+    module: 'auth',
+}) => {
+    const newPermissions = await getNewPermissions(dataSource, permissionOptions.table, permissionOptions.column);
+
+    if (!newPermissions.length) {
+        // eslint-disable-next-line no-console
+        console.log('info', 'No changes in permissions were found');
+        return;
+    }
+
+    const cliConfiguration = await loadConfiguration();
+    const dirPath = join(process.cwd(), cliConfiguration.sourceRoot, permissionOptions.module, 'infrastructure', 'migrations');
+
+    const values = newPermissions
+        .map(key => `('${key}')`)
+        .join(',\n            ');
+
+    const upRaw = `INSERT INTO ${permissionOptions.table} (${permissionOptions.column}) VALUES\n    ${values};`;
+
+    const downRaw = `DELETE FROM ${permissionOptions.table} WHERE ${permissionOptions.column} IN (${newPermissions.map(permission => `'${permission}'`).join(', ')});`;
+
+    const upQueries = [
+        `        await queryRunner.query(\`${prettifyQuery(upRaw)}\`);`,
+    ];
+
+    const downQueries = [
+        `        await queryRunner.query(\`${prettifyQuery(downRaw)}\`);`,
+    ];
+
+    const timestamp = new Date().getTime();
+
+    const migrationFileContent = getTemplate(
+        ADD_PERMISSIONS_NAME,
+        timestamp,
+        upQueries,
+        downQueries,
+    );
+    const migrationFilePath = join(dirPath, `${timestamp}-${ADD_PERMISSIONS_NAME}.ts`);
+    // eslint-disable-next-line no-console
+    console.log('info', '\t' + migrationFilePath);
+    await CommandUtils.createFile(migrationFilePath, migrationFileContent);
+};
 
 export const generate = async (connection: Connection) => {
+    const hasPendingMigrations = await connection.showMigrations();
+
+    if (hasPendingMigrations) {
+        console.error('[ERROR!] Unapplied migrations detected. Database schema is out of sync.');
+        return;
+    }
+
     // Get mapping model name to table name
     const junctionTablesMap = {};
 
@@ -77,8 +132,7 @@ export const generate = async (connection: Connection) => {
                 glob(moduleDir + '/**/tables/*Table{.ts,.js}', (err, matches) => {
                     if (err) {
                         reject(err);
-                    }
-                    else {
+                    } else {
                         resolve(matches);
                     }
                 });
@@ -117,13 +171,13 @@ export const generate = async (connection: Connection) => {
             };
         }
         migrationsByTables[tableName].upQueries.push(
-            '        await queryRunner.query(`' + prettifyQuery(item.query.query) + '`' + queryParams(item.query.parameters) + ');'
+            '        await queryRunner.query(`' + prettifyQuery(item.query.query) + '`' + queryParams(item.query.parameters) + ');',
         );
     }
     for (const item of sqlInMemory.downTableQueries) {
         const tableName = junctionTablesMap[item.tableName] || item.tableName;
         migrationsByTables[tableName].downQueries.push(
-            '        await queryRunner.query(`' + prettifyQuery(item.query.query) + '`' + queryParams(item.query.parameters) + ');'
+            '        await queryRunner.query(`' + prettifyQuery(item.query.query) + '`' + queryParams(item.query.parameters) + ');',
         );
     }
 
@@ -144,10 +198,8 @@ export const generate = async (connection: Connection) => {
             const tablePath = tablesInfo[tableName].tablePath;
 
             //Создание файла миграции с созданием таблиц, если такие запросы существуют
-            const tableDeclarationUpQueries =
-                migrationsByTables[tableName].upQueries.filter(query => query.includes('CREATE TABLE'));
-            const tableDeclarationDownQueries =
-                migrationsByTables[tableName].downQueries.filter(query => query.includes('DROP TABLE')).reverse();
+            const tableDeclarationUpQueries = migrationsByTables[tableName].upQueries.filter(query => query.includes('CREATE TABLE'));
+            const tableDeclarationDownQueries = migrationsByTables[tableName].downQueries.filter(query => query.includes('DROP TABLE')).reverse();
 
             if (tableDeclarationUpQueries.length > 0 || tableDeclarationDownQueries.length > 0) {
                 const tableDeclarationFileContent = getTemplate(
@@ -167,10 +219,8 @@ export const generate = async (connection: Connection) => {
             //Создание файла миграции с добавлением в таблицу внешних ключей, если такие запросы существуют
             const nextTimestamp = timestamp + 1;
 
-            const foreignKeysUpQueries =
-                migrationsByTables[tableName].upQueries.filter(query => !query.includes('CREATE TABLE'));
-            const foreignKeysDownQueries =
-                migrationsByTables[tableName].downQueries.filter(query => !query.includes('DROP TABLE')).reverse();
+            const foreignKeysUpQueries = migrationsByTables[tableName].upQueries.filter(query => !query.includes('CREATE TABLE'));
+            const foreignKeysDownQueries = migrationsByTables[tableName].downQueries.filter(query => !query.includes('DROP TABLE')).reverse();
 
             if (foreignKeysUpQueries.length > 0 || foreignKeysDownQueries.length > 0) {
                 const foreignKeysFileContent = getTemplate(
@@ -188,4 +238,4 @@ export const generate = async (connection: Connection) => {
             }
         }
     }
-}
+};