+++ src/infrastructure/decorators/typeorm/fields/TypeOrmDecimalNumberField/index.ts
@@ -1,5 +1,6 @@
 import {Column} from '@steroidsjs/typeorm';
 import {IDecimalFieldOptions} from '../../../fields/DecimalField';
+import {DEFAULT_DECIMAL_SCALE} from '../../../../base/consts';
 
 export default (options: IDecimalFieldOptions) => [
     Column({
@@ -7,7 +8,7 @@ export default (options: IDecimalFieldOptions) => [
         default: options.defaultValue,
         nullable: options.nullable,
         precision: options.precision || 10,
-        scale: options.scale || 2,
+        scale: options.scale || DEFAULT_DECIMAL_SCALE,
         array: options.isArray,
     }),
 ];